requirements.txt
nose
flake8
flake8-import-order
termcolor
xmltodict
gitpython
rosinstall_generator
rosdistro
requests==2.28.1
urllib3==1.26.12
rosdep
pep8
docker
pyyaml
pygithub
catkin_pkg
rospkg
